Fix price_change validation to accept float values from Deliverect
- Changed price_change from int to float in MenuModifierBase and MenuModifierUpdate
- Deliverect sends decimal prices like 0.5, not integer cents
- This resolves the Pydantic validation error blocking menu updates

+++ app/schemas/menu.py
@@ -154,7 +154,7 @@ class MenuItemUpdate(BaseModel):
 class MenuModifierBase(BaseModel):
     name: str
     description: Optional[str] = None
-    price_change: int = 0  # Price change in cents
+    price_change: float = 0.0  # Price change (can be decimal)
     plu: Optional[str] = None  # Changed to optional to match model
     deliverect_modifier_id: Optional[str] = None  # Added missing field
     location_id: Optional[str] = None  # Added missing field
@@ -169,7 +169,7 @@ class MenuModifierCreate(MenuModifierBase):
 class MenuModifierUpdate(BaseModel):
     name: Optional[str] = None
     description: Optional[str] = None
-    price_change: Optional[int] = None
+    price_change: Optional[float] = None
     plu: Optional[str] = None
     deliverect_modifier_id: Optional[str] = None  # Added missing field
     location_id: Optional[str] = None  # Added missing field
